requirements.txt
environs==14.1.0
Flask==3.1.0
protobuf==5.29.3
pydantic==2.10.5
httpx==0.28.1
pyngrok==7.2.3
python-dotenv==1.0.1
Requests==2.32.3
grpcio-tools==1.69.0
mypy-protobuf==3.6.0
pytest==8.3.4
pytest-mock==3.14.0